fix(product): keep the sdk-react query-key modules under the size threshold
PR5b's v2 key additions pushed both shared query-keys.ts files past the
600-line max (610 and 640): the workflow-run and workflows-v2 key
functions move to their own modules, re-exported from each package index
so consumers are unchanged. The anyharness client core.ts facade grows
by exactly the three registration lines every subsystem client adds; its
allowlist entry is bumped 603 -> 606 with that reason.
