+++ cloud/sdk-react/src/index.ts
@@ -17,3 +17,4 @@ export * from "./hooks/workspaces.js";
 export * from "./hooks/workflows.js";
 export * from "./hooks/workflows-v2.js";
 export * from "./lib/query-keys.js";
+export * from "./lib/query-keys-workflows-v2.js";
