+++ scripts/max_lines_allowlist.txt
@@ -33,6 +33,7 @@ anyharness/crates/anyharness-lib/src/live/terminals/manager_tests.rs 730 R3 work
 anyharness/crates/anyharness-lib/src/persistence/custom_migrations.rs 1318 custom migration with legacy schema coverage (+0063 partial unique controller index, ordered after the 0062 rebuild) ; +110: R1 archiving adds the 0067_workspace_archived_lifecycle FK-off full-table rebuild ; +72: R5 adds the 0068_workspace_drop_cleanup_columns FK-off full-table rebuild ; +10: workflows gen-2 PR1 registers the 0069_workflow_runs_gen2 FK-off rebuild
 anyharness/crates/proliferate-diagnostics-client/src/fallback/tests.rs 736 Fallback writer suite: every test drives one writer through the same rotation and live-authority harness (injected write errno, short writes, rotation, retirement, consecutive-failure retirement), so the shared harness is the reason it is one file and the tests themselves are independent. Split into fallback/tests/{write,rotation,authority}.rs when the next injection seam lands.
 anyharness/crates/proliferate-diagnostics-client/src/fallback/unix.rs 651 Unix fallback writer: segment rotation, live-authority validation, partial-line rollback, and bounded interrupt-retrying writes all mutate the same file handle and byte counters under one lock, so the retry path cannot move without exporting that state. The cross-platform swap primitive already moved out to fallback/swap.rs; split the authority validation out next when a second platform needs it.
+anyharness/sdk/src/client/core.ts 606 AnyHarness TS client core (+preserved FastAPI detail body on AnyHarnessError for gateway billing-block classification, C8) ; +3: gen-2 workflow-runs client registration (import + facade field + constructor wiring, the same three lines every subsystem client adds)
 anyharness/sdk-react/src/hooks/sessions.ts 675 existing SDK React oversized file (subagent queries and lifecycle mutations extracted to their owned hook module)
 apps/desktop/src-tauri/src/diagnostics/support_snapshot/artifact_store/unix_fs.rs 658 Support snapshot artifact store's Unix filesystem layer: private-by-construction root creation, metadata repair, and the O_NOFOLLOW open/read/write paths share the same permission constants and repair state machine, and the mkdir-mode hardening this slice adds is only meaningful next to the repair it guards. Split into unix_fs/{root,entry}.rs when a second artifact kind lands.
 apps/desktop/src-tauri/src/diagnostics/support_snapshot/scrub/patterns.rs 610 Support export scrub pattern table: the labelled, unlabelled-scheme, environment-assignment, provider-prefix, and opaque-catch-all rules form one ordered redaction pipeline whose precedence is load-bearing and pinned by the scrub tests, so a rule family cannot move without moving that order with it. Split into scrub/patterns/{labelled,unlabelled,opaque}.rs when a sixth rule family lands.
